requirements.txt
boto3==1.28.0  # AWS SDK for Python
google-cloud-storage==2.10.0  # Google Cloud Storage client library
google-auth==2.21.0  # Authentication for Google APIs
requests==2.31.0  # For making HTTP requests
python-dotenv==1.0.0
